requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
redis[hiredis]==5.0.1
pydantic[email]==2.5.1
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.23.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
PyJWT==2.8.0
numpy>=1.26.4
httpx==0.25.2
torch==2.8.0
gymnasium==1.2.0
stable-baselines3==2.7.0
faiss-cpu==1.12.0
sentence-transformers==3.4.0
